an edit of script and persistence model for tagging of persistence
- Controllers.kt containing all controllers
