+++ persistence/src/test/kotlin/com/github/jactor/persistence/AbstractSpringBootNoDirtyContextTest.kt
@@ -10,12 +10,12 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import com.github.jactor.persistence.address.AddressRepository
 import com.github.jactor.persistence.blog.BlogEntryRepository
 import com.github.jactor.persistence.blog.BlogRepository
+import com.github.jactor.persistence.blog.BlogService
 import com.github.jactor.persistence.guestbook.GuestBookEntryRepository
 import com.github.jactor.persistence.guestbook.GuestBookRepository
+import com.github.jactor.persistence.guestbook.GuestBookService
 import com.github.jactor.persistence.person.PersonRepository
 import com.github.jactor.persistence.user.UserRepository
-import com.github.jactor.persistence.blog.BlogService
-import com.github.jactor.persistence.guestbook.GuestBookService
 import com.ninjasquad.springmockk.SpykBean
 import jakarta.persistence.EntityManager
 
@@ -78,13 +78,14 @@ abstract class AbstractSpringBootNoDirtyContextTest {
     @Autowired
     private lateinit var entityManager: EntityManager
 
-    protected fun<T> flush(databaseOperation: () -> T): T {
+    protected fun <T> flush(databaseOperation: () -> T): T {
         val entity = databaseOperation.invoke()
         entityManager.flush()
         entityManager.clear()
 
         return entity
     }
+
     protected fun buildFullPath(url: String): String {
         return "http://localhost:$port$contextPath$url"
     }
